playback: improve persistence
Improve playback persistence in the following ways:
1. Shift the boundary of PlaybackStateManager and PlaybackStateDatabase
so that the reading and searching phases both occur at the same time,
which is more efficient.
2. Improve music hashing so that conflicts are minimized [this also
helps the future playlists addition]
3. Generally improve code style

+++ app/src/main/java/org/oxycblt/auxio/music/MusicStore.kt
@@ -92,7 +92,7 @@ class MusicStore private constructor() {
     /**
      * Find a song in a faster manner using a hash for its album as well.
      */
-    fun findSongFast(songHash: Int, albumHash: Int): Song? {
+    fun findSongFast(songHash: Long, albumHash: Long): Song? {
         return albums.find { it.hash == albumHash }?.songs?.find { it.hash == songHash }
     }
